+++ Mockingbird/Models/Endpoint.cs
@@ -1,6 +1,4 @@
 using System.Collections.Generic;
-using System.Net.Http;
-using Microsoft.AspNetCore.Http;
 
 namespace Mockingbird.Models
 {
@@ -8,8 +6,8 @@ namespace Mockingbird.Models
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public MockingbirdRequest Request { get; set; }
-        public MockingbirdResponse Response { get; set; }
+        public EndpointRequest Request { get; set; }
+        public EndpointResponse Response { get; set; }
         public IEnumerable<Directory> Directory { get; set; } = new List<Directory>();
         public IEnumerable<Tag> Tags { get; set; } = new List<Tag>();
         public bool Enabled { get; set; }
