+++ Mockingbird/Models/EndpointRequest.cs
@@ -2,7 +2,7 @@ using System.Net.Http;
 
 namespace Mockingbird.Models
 {
-    public sealed class MockingbirdRequest
+    public sealed class EndpointRequest
     {
         public HttpMethod Method { get; set; }
         public string PathMatcherPattern { get; set; }
